refactor: split KsqlContext
