+++ docs/diff_log/diff_ksqlcontext_refactor_20250909.md
@@ -1,0 +1,2 @@
+# diff_ksqlcontext_refactor_20250909
+- Split KsqlContext into partial classes for entity management and schema registration to reduce file size.
