+++ features/ksqlcontext_refactor/instruction.md
@@ -1,0 +1,3 @@
+# ksqlcontext_refactor
+
+KsqlContext.cs has grown large. Split entity management and schema registration logic into separate partial class files to reduce size.
